+++ src/main/java/racingcar/model/Car.java
@@ -2,7 +2,7 @@ package racingcar.model;
 
 import racingcar.model.wrapper.Name;
 import racingcar.model.wrapper.Position;
-import racingcar.utils.RacingNumberGenerator;
+import racingcar.strategy.RacingNumberGenerator;
 
 public class Car {
