+++ src/main/java/racingcar/model/wrapper/Position.java
@@ -1,6 +1,6 @@
 package racingcar.model.wrapper;
 
-import racingcar.utils.RacingNumberGenerator;
+import racingcar.strategy.RacingNumberGenerator;
 
 public class Position implements Comparable<Position> {
